// frontend/src/shared/api/generated/model/sessionDetailRead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { FeynmanBlockRead } from './feynmanBlockRead';
import type { FeynmanMessageRead } from './feynmanMessageRead';
import type { FeynmanSessionRead } from './feynmanSessionRead';

export interface SessionDetailRead {
  session: FeynmanSessionRead;
  feynman_block: FeynmanBlockRead;
  messages: FeynmanMessageRead[];
}